+++ packages/camera/camera_platform_interface/lib/src/types/camera_description.dart
@@ -2,6 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/foundation.dart';
+import 'package:meta/meta.dart';
+
 /// The direction the camera is facing.
 enum CameraLensDirection {
   /// Front facing camera (a user looking at the screen is seen by the camera).
@@ -15,9 +18,10 @@ enum CameraLensDirection {
 }
 
 /// Properties of a camera device.
+@immutable
 class CameraDescription {
   /// Creates a new camera description with the given properties.
-  CameraDescription({
+  const CameraDescription({
     required this.name,
     required this.lensDirection,
     required this.sensorOrientation,
@@ -51,6 +55,7 @@ class CameraDescription {
 
   @override
   String toString() {
-    return '$runtimeType($name, $lensDirection, $sensorOrientation)';
+    return '${objectRuntimeType(this, 'CameraDescription')}('
+        '$name, $lensDirection, $sensorOrientation)';
   }
 }