+++ packages/webview_flutter/webview_flutter_android/pigeons/android_webview.dart
@@ -18,16 +18,46 @@ class WebResourceErrorData {
   String? description;
 }
 
+@HostApi()
+abstract class CookieManagerHostApi {
+  @async
+  bool clearCookies();
+
+  void setCookie(String url, String value);
+}
+
 @HostApi(dartHostTestHandler: 'TestWebViewHostApi')
 abstract class WebViewHostApi {
   void create(int instanceId, bool useHybridComposition);
 
   void dispose(int instanceId);
 
+  void loadData(
+    int instanceId,
+    String data,
+    String mimeType,
+    String encoding,
+  );
+
+  void loadDataWithBaseUrl(
+    int instanceId,
+    String baseUrl,
+    String data,
+    String mimeType,
+    String encoding,
+    String historyUrl,
+  );
+
   void loadUrl(
     int instanceId,
     String url,
     Map<String, String> headers,
+  );
+
+  void postUrl(
+    int instanceId,
+    String url,
+    Uint8List data,
   );
 
   String getUrl(int instanceId);
@@ -71,6 +101,8 @@ abstract class WebViewHostApi {
   void setDownloadListener(int instanceId, int listenerInstanceId);
 
   void setWebChromeClient(int instanceId, int clientInstanceId);
+
+  void setBackgroundColor(int instanceId, int color);
 }
 
 @HostApi(dartHostTestHandler: 'TestWebSettingsHostApi')
@@ -100,29 +132,31 @@ abstract class WebSettingsHostApi {
   void setDisplayZoomControls(int instanceId, bool enabled);
 
   void setBuiltInZoomControls(int instanceId, bool enabled);
+
+  void setAllowFileAccess(int instanceId, bool enabled);
 }
 
 @HostApi(dartHostTestHandler: 'TestJavaScriptChannelHostApi')
 abstract class JavaScriptChannelHostApi {
   void create(int instanceId, String channelName);
-
-  void dispose(int instanceId);
 }
 
 @FlutterApi()
 abstract class JavaScriptChannelFlutterApi {
+  void dispose(int instanceId);
+
   void postMessage(int instanceId, String message);
 }
 
 @HostApi(dartHostTestHandler: 'TestWebViewClientHostApi')
 abstract class WebViewClientHostApi {
   void create(int instanceId, bool shouldOverrideUrlLoading);
-
-  void dispose(int instanceId);
 }
 
 @FlutterApi()
 abstract class WebViewClientFlutterApi {
+  void dispose(int instanceId);
+
   void onPageStarted(int instanceId, int webViewInstanceId, String url);
 
   void onPageFinished(int instanceId, int webViewInstanceId, String url);
@@ -154,11 +188,12 @@ abstract class WebViewClientFlutterApi {
 @HostApi(dartHostTestHandler: 'TestDownloadListenerHostApi')
 abstract class DownloadListenerHostApi {
   void create(int instanceId);
-  void dispose(int instanceId);
 }
 
 @FlutterApi()
 abstract class DownloadListenerFlutterApi {
+  void dispose(int instanceId);
+
   void onDownloadStart(
     int instanceId,
     String url,
@@ -172,10 +207,18 @@ abstract class DownloadListenerFlutterApi {
 @HostApi(dartHostTestHandler: 'TestWebChromeClientHostApi')
 abstract class WebChromeClientHostApi {
   void create(int instanceId, int webViewClientInstanceId);
-  void dispose(int instanceId);
+}
+
+@HostApi(dartHostTestHandler: 'TestAssetManagerHostApi')
+abstract class FlutterAssetManagerHostApi {
+  List<String> list(String path);
+
+  String getAssetFilePathByName(String name);
 }
 
 @FlutterApi()
 abstract class WebChromeClientFlutterApi {
+  void dispose(int instanceId);
+
   void onProgressChanged(int instanceId, int webViewInstanceId, int progress);
 }